+++ typescript/packages/server-express/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @bosonprotocol/x402-server-express
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies [731acc3]
+  - @bosonprotocol/x402-server@0.3.0
+
 ## 0.1.2
 
 ### Patch Changes